+++ src/entities/ActionsPanel/ActionsPanel.tsx
@@ -11,10 +11,19 @@ import { IconButton } from '@shared/ui/molecules';
 import { RIGHT_PANEL_TYPE } from '@shared/constants';
 import { useDebouncedCallback } from '@src/shared/hooks/useDebouncedCallback';
 import { useDeleteRightModelPanelStore } from '@src/shared/stores';
-import { useRoles } from '@src/shared/hooks';
+import { usePermissions, useRoles } from '@src/shared/hooks';
 
 import { CustomSearchInput, Container } from './styles';
 import { ROUTES } from '../../app/Routes';
+
+// constants.ts
+export enum DELETE_TOOLTIP_MESSAGES {
+  NO_MODEL_SELECTED = 'Выберите модель для удаления',
+  MULTIPLE_MODELS_SELECTED = 'Нельзя удалить несколько моделей',
+  INVALID_SOURCE = 'Модель должна быть с источником "sum-rm"',
+  NOT_AUTHORIZED = 'Модель может быть удалена только создателем, владельцем модели или администратором',
+  DELETE_MODEL = 'Удалить модель',
+}
 
 export interface ActionsPanelProps {
   handleSearch: (newSearchString: string) => void;
@@ -23,24 +32,18 @@ export interface ActionsPanelProps {
 
 export const ActionsPanel = ({ updateRightPanelType, handleSearch }: ActionsPanelProps) => {
   const [searchValue, setSearchValue] = useState('');
-  const { modelsCount, modelSource, isDeleteButtonEnabled, userMatches, modelStatus } =
+  const { modelsCount, modelSource, isDeleteButtonEnabled, userMatches } =
     useDeleteRightModelPanelStore();
   const { isAdmin, isValidatorLead } = useRoles();
+  const { isAddModelEnabled } = usePermissions();
 
-  let deleteTooltipMessage = '';
-
-  if (modelsCount === 0) {
-    deleteTooltipMessage = 'Выберите модель для удаления';
-  } else if (modelsCount > 1) {
-    deleteTooltipMessage = 'Нельзя удалить несколько моделей';
-  } else if (modelSource !== 'sum-rm') {
-    deleteTooltipMessage = 'Модель должна быть с исчтоником "sum-rm"';
-  } else if (!userMatches && !isAdmin && !isValidatorLead) {
-    deleteTooltipMessage =
-      'Модель может-быть удалена только создателем, владельцем модели или администратором';
-  } else {
-    deleteTooltipMessage = 'Удалить модель';
-  }
+  const deleteTooltipMessage = (() => {
+    if (modelsCount === 0) return DELETE_TOOLTIP_MESSAGES.NO_MODEL_SELECTED;
+    if (modelsCount > 1) return DELETE_TOOLTIP_MESSAGES.MULTIPLE_MODELS_SELECTED;
+    if (modelSource !== 'sum-rm') return DELETE_TOOLTIP_MESSAGES.INVALID_SOURCE;
+    if (!userMatches && !isAdmin && !isValidatorLead) return DELETE_TOOLTIP_MESSAGES.NOT_AUTHORIZED;
+    return DELETE_TOOLTIP_MESSAGES.DELETE_MODEL;
+  })();
 
   const debouncedHandleChange = useDebouncedCallback((e: React.ChangeEvent<HTMLInputElement>) => {
     handleChange(e);
@@ -65,19 +68,21 @@ export const ActionsPanel = ({ updateRightPanelType, handleSearch }: ActionsPane
         value={searchValue}
       />
       <div>
+        {isAddModelEnabled && (
+          <IconButton
+            icon={<PlusCircleSolid />}
+            tooltip="Добавить модель"
+            color="#0062FF"
+            onClick={() => updateRightPanelType(RIGHT_PANEL_TYPE.ADD_MODEL)}
+          />
+        )}
         <IconButton
-          icon={<PlusCircleSolid />}
-          tooltip="Добавить модель"
-          color="#0062FF"
-          onClick={() => updateRightPanelType(RIGHT_PANEL_TYPE.ADD_MODEL)}
-        />
-        {/* <IconButton
           icon={<DeleteSolid />}
           tooltip={deleteTooltipMessage}
           color="#0062FF"
           onClick={() => updateRightPanelType(RIGHT_PANEL_TYPE.DELETE_MODEL)}
           disabled={!isDeleteButtonEnabled}
-        /> */}
+        />
 
         <IconButton
           icon={<BrokerOutlineIcon />}