+++ src/features/RightModalPanel/ModelForm/ModelForm.tsx
@@ -9,13 +9,13 @@ import { StatusScreen } from '@shared/ui/molecules';
 import { RIGHT_PANEL_TYPE, MODEL_FORM_MODE } from '@shared/constants';
 import { INPUT_TYPE, InputFactory, InputValue, RightPanel } from '@shared/ui/organisms';
 import { API_ROUTES, useFetch, ArtifactApi, ModelEditApi } from '@shared/api';
+import { useModelUserMatch, useRoles } from '@src/shared/hooks';
 
-import { groupBy, isEqual, omit, pick, uniqBy } from 'lodash';
+import { groupBy, isEqual, omit, uniqBy } from 'lodash';
 import { Flexbox, Spacer } from '@shared/ui/atoms';
 import { Artifact } from '@shared/api/types';
 
 import { useAppInjectStore } from '@shared/stores/appInjectStore';
-import { CUSTOMER_MAP } from '@shared/constants/customers';
 import { useScrollTo } from '@src/shared/hooks/useScrollTo';
 import { useDeepEffect } from '@src/shared/hooks/useDeepEffect';
 import { FormValues } from '../types';
@@ -74,6 +74,11 @@ export const ModelForm = ({
     { connectedName: string; connectedValue: string }[] | undefined
   >(undefined);
 
+  const { isValidator, isValidatorLead } = useRoles();
+  const { isInBusinessCustomers } = useModelUserMatch();
+
+  const isEditByRatingModel = isValidator || isValidatorLead || isInBusinessCustomers(activeRow!);
+
   const errorElemRef = useRef<HTMLDivElement>(null);
   const formRef = useRef<HTMLFormElement | null>(null);
 
@@ -102,7 +107,7 @@ export const ModelForm = ({
 
   const IS_FORM_MODE_ADD = formMode === MODEL_FORM_MODE.ADD;
   const title = IS_FORM_MODE_ADD ? 'Новая модель' : 'Редактирование модели';
-  const groupedBySchemaName = groupBy(fields, 'schemaKey');
+  const groupedFieldsBySchemaName = groupBy(fields, 'schemaKey');
 
   const handleChange = (name: keyof Row, value: InputValue) => {
     const newValues = { [name]: value };
@@ -355,7 +360,7 @@ export const ModelForm = ({
         setSubmitError('');
       }
     },
-    [values, formSchema, formMode, activeModelByDefault],
+    [values, formSchema, formMode, activeModelByDefault, fields],
   );
 
   const handleChangeParentModel = useCallback(
@@ -386,7 +391,6 @@ export const ModelForm = ({
 
   const activeModelCheckboxHandler = async (e: any) => {
     setActiveModelByDefault(e?.target?.checked);
-    setCurrentCustomer(CUSTOMER_MAP.UMRV);
 
     await handleSubmit({ checkOnly: true });
   };
@@ -398,19 +402,30 @@ export const ModelForm = ({
   }, [initialRow, artifacts]);
 
   useEffect(() => {
+    // if (!isEditByRatingModel) {
+    //   setActiveModelByDefault(false);
+    //   return;
+    // }
+
     if (activeRow?.active_model === '1') {
       setActiveModelByDefault(true);
     }
-  }, [activeRow?.active_model]);
+  }, [activeRow?.active_model, isEditByRatingModel]);
 
   // Scroll to edit input field
   useEffect(() => {
     if (formRef.current?.children && editCellName) {
-      const editedFieldIndex = fields.findIndex((field) => field.name === editCellName);
+      const inputFieldsRefs = document.querySelectorAll(
+        '#model_form_parent_container [data-form-input]',
+      );
+
+      const editedFieldIndex = Array.from(inputFieldsRefs).findIndex(
+        (field) => field.getAttribute('data-form-input') === editCellName,
+      );
 
       if (editedFieldIndex !== -1) {
-        formRef.current.children[editedFieldIndex]?.scrollIntoView({
-          block: 'center',
+        Array.from(inputFieldsRefs)[editedFieldIndex]?.scrollIntoView({
+          block: 'start',
           behavior: 'smooth',
         });
       }
@@ -459,6 +474,7 @@ export const ModelForm = ({
               dimension="s"
               checked={activeModelByDefault}
               onChange={activeModelCheckboxHandler}
+              // disabled={!isEditByRatingModel}
             >
               Действующая Модель/Модуль
             </CheckboxField>
@@ -473,9 +489,9 @@ export const ModelForm = ({
           apiLoading={submitLoading}
           onFinished={handleOnClose}
         >
-          <FormContainer ref={formRef}>
-            {Object.keys(groupedBySchemaName).map((schemaKey) => {
-              const fieldsByGroup = groupedBySchemaName[schemaKey || 'Аллокация'];
+          <FormContainer ref={formRef} id="model_form_parent_container">
+            {Object.keys(groupedFieldsBySchemaName).map((schemaKey) => {
+              const fieldsByGroup = groupedFieldsBySchemaName[schemaKey || 'Аллокация'];
               const schemaTitle = SCHEMA_NAME_MAP[schemaKey]?.title;
 
               return (
@@ -499,6 +515,7 @@ export const ModelForm = ({
                           fillChild
                           key={field.name}
                           ref={errorElemRef}
+                          data-form-input={field.name}
                         >
                           <InputFactory<keyof Row>
                             values={values}