+++ src/pages/Playground/PlaygroundTable.tsx
@@ -35,11 +35,12 @@ import { useAppInjectStore } from '@src/shared/stores/appInjectStore';
 import { CUSTOMER_MAP } from '@src/shared/constants/customers';
 import { useTableChange } from '@src/features/Tables/hooks';
 import { format } from 'date-fns';
+import { Template } from '@src/shared/api/types';
 import { PlaygroundCustomCell } from './PlaygroundCustomCell';
 import { AG_GRID_LOCALE_RU } from './locale/agGridLocale.ru';
 import { ROUTES } from '../../app/Routes';
 import { useDeleteRightModelPanelStore } from '../../shared/stores';
-import { useRoles, useTemplateFilters } from '../../shared/hooks';
+import { usePermissions, useRoles, useTemplateFilters } from '../../shared/hooks';
 import { isInBusinessCustomers, isModelCreator } from '../../shared/helpers';
 import { useDeepEffect } from '../../shared/hooks/useDeepEffect';
 
@@ -75,12 +76,15 @@ export const PlaygroundTable = ({
   display,
   modelsTable,
   filters,
+  templates,
 }: {
   display: TDisplayTableModels;
   modelsTable: TModelsTable;
   filters: TFilters;
+  templates: Template[];
 }) => {
   const { rowList, setPage, page, setTotalRows, pageSize, searchString, columnList } = modelsTable;
+
   const {
     cols,
     rows,
@@ -115,6 +119,7 @@ export const PlaygroundTable = ({
     topFilters?.templates,
   );
   const { isAdmin, isValidatorLead } = useRoles();
+  const { isAddModelEnabled } = usePermissions();
 
   const navigate = useNavigate();
   const gridRef = useRef<AgGridReact>(null);
@@ -123,6 +128,7 @@ export const PlaygroundTable = ({
     ...data,
     headerName: data.title,
     field: data.name,
+    headerTooltip: data.title,
     // https://www.ag-grid.com/react-data-grid/filter-date/#custom-selection-component
     filter:
       data.type === COLUMN_TYPE.DATE
@@ -139,7 +145,7 @@ export const PlaygroundTable = ({
       filter: 'agMultiColumnFilter',
       floatingFilter: true,
       initialWidth: 400,
-      minWidth: 200,
+      minWidth: 250,
       maxWidth: 1350,
       suppressHeaderMenuButton: false,
       suppressHeaderContextMenu: false,
@@ -152,8 +158,8 @@ export const PlaygroundTable = ({
       flex: 2,
       sortable: true,
       resizable: true,
-      wrapHeaderText: true,
-      autoHeaderHeight: true,
+      wrapHeaderText: false,
+      autoHeaderHeight: false,
       // valueGetter: (params: ValueGetterParams) => {
       //   return `(${params.getValue})`;
       // },
@@ -262,7 +268,7 @@ export const PlaygroundTable = ({
       setRows(rowList);
       modelsTable.setTotalRows(rowList.length);
     }
-  }, [rowList, modelsTable.setTotalRows]);
+  }, [rowList, modelsTable.setTotalRows, templates]);
 
   return (
     <Flexbox height="calc(100vh - 230px)">
@@ -277,12 +283,14 @@ export const PlaygroundTable = ({
             />
           </Flexbox>
           <div>
-            <IconButton
-              icon={<PlusCircleSolid />}
-              tooltip="Добавить модель"
-              color="#0062FF"
-              onClick={() => display.setRightPanelType(RIGHT_PANEL_TYPE.ADD_MODEL)}
-            />
+            {isAddModelEnabled && (
+              <IconButton
+                icon={<PlusCircleSolid />}
+                tooltip="Добавить модель"
+                color="#0062FF"
+                onClick={() => display.setRightPanelType(RIGHT_PANEL_TYPE.ADD_MODEL)}
+              />
+            )}
             {/* <IconButton
               icon={<DeleteSolid />}
               tooltip={deleteTooltipMessage}
@@ -349,11 +357,12 @@ export const PlaygroundTable = ({
             onSelectionChanged={handleSelectionChange}
             onFilterChanged={handleFilterChange}
             pagination
-            paginationPageSize={100}
+            paginationPageSize={pageSize}
             paginationPageSizeSelector={paginationPageSizeSelector}
             singleClickEdit
             localeText={AG_GRID_LOCALE_RU}
             alwaysShowHorizontalScroll
+            tooltipShowDelay={500}
           />
         </div>
       </div>