+++ src/shared/hooks/usePermissions.ts
@@ -1,0 +1,14 @@
+import { useUserStore } from '../stores';
+import { Permission, Role } from '../types';
+
+export const usePermissions = () => {
+  const { permissions, hasPermission } = useUserStore();
+
+  return {
+    permissions,
+    hasPermission,
+    isAddModelEnabled: hasPermission(Permission.ADD_MODEL),
+    isEditModelEnabled: hasPermission(Permission.EDIT_MODEL),
+    isAddPublicTemplateEnabled: hasPermission(Permission.ADD_PUBLIC_TEMPLATE),
+  };
+};