+++ src/shared/stores/index.ts
@@ -1,4 +1,5 @@
 export { useAppInjectStore } from './appInjectStore';
 export { useDeleteRightModelPanelStore } from './deleteRightModelPanelStore';
 export { useUserStore } from './userStore';
+export { useExcludeErrorStore } from './excludeErrorStore';