+++ src/shared/types/index.ts
@@ -1,3 +1,4 @@
 export * from './general';
 export * from './roles';
+export * from './permissions';