+++ src/shared/ui/atoms/Pagination/Pagination.tsx
@@ -1,6 +1,8 @@
 import * as React from 'react';
-import { PaginationOne } from '@admiral-ds/react-ui';
 import styled from 'styled-components';
+
+import { Checkbox, PaginationOne, T } from '@admiral-ds/react-ui';
+import { useExcludeErrorStore } from '@src/shared/stores/excludeErrorStore';
 
 export interface PaginationProps {
   page: number;
@@ -11,9 +13,27 @@ export interface PaginationProps {
 }
 
 const CustomPagination = styled(PaginationOne)`
+  box-sizing: border-box;
+`;
+
+const Container = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
   background: var(--neutral-neutral-05, #f3f4f6);
   padding: 8px;
-  box-sizing: bored-box;
+`;
+
+const Row = styled.div`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  > *:not(:last-child) {
+    margin-right: 6px;
+  }
+  > * {
+    flex: 0 0 auto;
+  }
 `;
 
 export const Pagination = ({
@@ -22,12 +42,28 @@ export const Pagination = ({
   totalElements,
   pageSizes = [10, 20, 50, 100, 200],
   onChangePage,
-}: PaginationProps) => (
-  <CustomPagination
-    page={page}
-    pageSize={pageSize}
-    totalItems={totalElements}
-    pageSizes={pageSizes}
-    onChange={onChangePage}
-  />
-);
+}: PaginationProps) => {
+  const { excludeError, updateExcludeError } = useExcludeErrorStore();
+
+  return (
+    <Container>
+      <Row>
+        <Checkbox
+          dimension="s"
+          checked={excludeError}
+          onChange={(e) => updateExcludeError(e.target.checked)}
+        />
+        <T font="Body/Body 2 Short" as="div">
+          Не включать модели со статусом ошибка заведения
+        </T>
+      </Row>
+      <CustomPagination
+        page={page}
+        pageSize={pageSize}
+        totalItems={totalElements}
+        pageSizes={pageSizes}
+        onChange={onChangePage}
+      />
+    </Container>
+  );
+};